feat(EDCO): Added rowMaxSteps and stepsAlignment props, updated layout and story
